+++ packages/expo-dev-launcher/android/src/react-native-72/expo/modules/devlauncher/rncompatibility/DevLauncherDevSupportManager.kt
@@ -1,6 +1,7 @@
 package expo.modules.devlauncher.rncompatibility
 
 import android.content.Context
+import android.net.Uri
 import android.os.Handler
 import android.os.Looper
 import android.util.Log
@@ -66,7 +67,8 @@ class DevLauncherDevSupportManager(
 
   // copied from https://github.com/facebook/react-native/blob/aa4da248c12e3ba41ecc9f1c547b21c208d9a15f/ReactAndroid/src/main/java/com/facebook/react/devsupport/BridgeDevSupportManager.java#L65
   private var mIsSamplingProfilerEnabled = false
-  private val devSettings: DevLauncherInternalSettingsWrapper = DevLauncherInternalSettingsWrapper(getDevSettings())
+  private val devSettings: DevLauncherInternalSettingsWrapper =
+    DevLauncherInternalSettingsWrapper(getDevSettings())
 
   // copied from https://github.com/facebook/react-native/blob/aa4da248c12e3ba41ecc9f1c547b21c208d9a15f/ReactAndroid/src/main/java/com/facebook/react/devsupport/BridgeDevSupportManager.java#L88-L128
   init {
@@ -99,7 +101,10 @@ class DevLauncherDevSupportManager(
 
   override fun showNewJavaError(message: String?, e: Throwable) {
     if (!DevLauncherController.wasInitialized()) {
-      Log.e("DevLauncher", "DevLauncher wasn't initialized. Couldn't intercept native error handling.")
+      Log.e(
+        "DevLauncher",
+        "DevLauncher wasn't initialized. Couldn't intercept native error handling."
+      )
       super.showNewJavaError(message, e)
       return
     }
@@ -126,7 +131,8 @@ class DevLauncherDevSupportManager(
       object : CallbackWithBundleLoader {
         override fun onSuccess(bundleLoader: JSBundleLoader) {
           bundleLoader.loadScript(currentContext!!.catalystInstance)
-          var bundleURL = controller?.manifest?.getBundleURL() ?: devServerHelper.getDevServerSplitBundleURL(bundlePath)
+          val bundleURL = controller?.manifest?.getBundleURL()
+            ?: devServerHelper.getDevServerSplitBundleURL(bundlePath)
           currentContext!!
             .getJSModule(HMRClient::class.java)
             .registerBundle(bundleURL)
@@ -224,6 +230,32 @@ class DevLauncherDevSupportManager(
         .getDevServerBundleURL(Assertions.assertNotNull(jsAppBundleName))
       reloadJSFromServer(bundleURL)
     }
+  }
+
+  override fun getSourceUrl(): String {
+    return controller?.manifest?.getBundleURL() ?: super.getSourceUrl()
+  }
+
+  override fun getSourceMapUrl(): String {
+    val defaultValue = super.getSourceMapUrl()
+    val bundleURL = controller?.manifest?.getBundleURL()
+      ?: return defaultValue
+
+    val parsedURL = Uri.parse(bundleURL)
+    val customOptions = parsedURL.queryParameterNames.mapNotNull { key ->
+      if (key.startsWith("transform")) {
+        key to requireNotNull(parsedURL.getQueryParameter(key))
+      } else {
+        null
+      }
+    }.ifEmpty {
+      return defaultValue
+    }
+
+    val customOptionsString = customOptions.joinToString("&") { (key, value) ->
+      "$key=$value"
+    }
+    return "$defaultValue&$customOptionsString"
   }
 
   // copied from https://github.com/facebook/react-native/blob/aa4da248c12e3ba41ecc9f1c547b21c208d9a15f/ReactAndroid/src/main/java/com/facebook/react/devsupport/BridgeDevSupportManager.java#L233-L277